rss_server/requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.1.7
django-cors-headers==2.5.2
djangorestframework==3.9.2
get==2018.11.19
idna==2.8
pep8==1.7.1
post==2018.11.20
public==2018.11.20
pytz==2018.9
query-string==2018.11.20
requests==2.21.0
urllib3==1.24.1
